Updated on 2026-02-02

+++ blockchain/src/main/java/com/tangem/blockchain/blockchains/cardano/network/adalite/converters/AdaliteUnspentOutputConverter.kt
@@ -27,7 +27,7 @@ internal object AdaliteUnspentOutputConverter {
             CardanoUnspentOutput.Asset(
                 policyID = token.policyId,
                 assetNameHex = token.assetName,
-                amount = token.quantity,
+                amount = token.quantity.toLong(),
             )
         }
     }
